dask-fallback/requirements.txt
dask[complete]==2024.1.0
distributed==2024.1.0
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
scikit-learn==1.4.0
influxdb==5.3.1
requests==2.31.0
avro-python3==1.10.2
pyyaml==6.0.1
